Get rid of any common classes
