+++ reactivesocket-transport-websocket/src/main/java/io/reactivesocket/transport/websocket/client/ClientWebSocketDuplexConnection.java
@@ -16,6 +16,8 @@
 package io.reactivesocket.transport.websocket.client;
 
 import io.netty.bootstrap.Bootstrap;
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.Unpooled;
 import io.netty.channel.*;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioSocketChannel;
@@ -23,20 +25,31 @@ import io.netty.handler.codec.http.DefaultHttpHeaders;
 import io.netty.handler.codec.http.HttpClientCodec;
 import io.netty.handler.codec.http.HttpObjectAggregator;
 import io.netty.handler.codec.http.websocketx.*;
+import io.reactivesocket.DuplexConnection;
 import io.reactivesocket.Frame;
-import io.reactivesocket.transport.tcp.NettyDuplexConnection;
+import io.reactivesocket.exceptions.TransportException;
+import io.reactivesocket.rx.Completable;
+import io.reactivesocket.rx.Observable;
 import io.reactivesocket.rx.Observer;
 import org.reactivestreams.Publisher;
+import org.reactivestreams.Subscriber;
+import org.reactivestreams.Subscription;
 
+import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.nio.channels.ClosedChannelException;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-public class ClientWebSocketDuplexConnection extends NettyDuplexConnection {
+public class ClientWebSocketDuplexConnection implements DuplexConnection {
+    private Channel channel;
 
-    private ClientWebSocketDuplexConnection(Channel channel, CopyOnWriteArrayList<Observer<Frame>> readers) {
-        super(channel, readers);
+    private final CopyOnWriteArrayList<Observer<Frame>> subjects;
+
+    private ClientWebSocketDuplexConnection(Channel channel, CopyOnWriteArrayList<Observer<Frame>> subjects) {
+        this.subjects  = subjects;
+        this.channel = channel;
     }
 
     public static Publisher<ClientWebSocketDuplexConnection> create(InetSocketAddress address, String path, EventLoopGroup eventLoopGroup) {
@@ -50,10 +63,10 @@ public class ClientWebSocketDuplexConnection extends NettyDuplexConnection {
     public static Publisher<ClientWebSocketDuplexConnection> create(URI uri, EventLoopGroup eventLoopGroup) {
         return s -> {
             WebSocketClientHandshaker handshaker = WebSocketClientHandshakerFactory.newHandshaker(
-                    uri, WebSocketVersion.V13, null, false, new DefaultHttpHeaders());
+                uri, WebSocketVersion.V13, null, false, new DefaultHttpHeaders());
 
-            CopyOnWriteArrayList<Observer<Frame>> readers = new CopyOnWriteArrayList<>();
-            ReactiveSocketClientHandler clientHandler = new ReactiveSocketClientHandler(readers);
+            CopyOnWriteArrayList<Observer<Frame>> subjects = new CopyOnWriteArrayList<>();
+            ReactiveSocketClientHandler clientHandler = new ReactiveSocketClientHandler(subjects);
             Bootstrap bootstrap = new Bootstrap();
             ChannelFuture connect = bootstrap
                 .group(eventLoopGroup)
@@ -78,7 +91,7 @@ public class ClientWebSocketDuplexConnection extends NettyDuplexConnection {
                         .getHandshakePromise()
                         .addListener(handshakeFuture -> {
                             if (handshakeFuture.isSuccess()) {
-                                s.onNext(new ClientWebSocketDuplexConnection(ch, readers));
+                                s.onNext(new ClientWebSocketDuplexConnection(ch, subjects));
                                 s.onComplete();
                             } else {
                                 s.onError(handshakeFuture.cause());
@@ -89,6 +102,71 @@ public class ClientWebSocketDuplexConnection extends NettyDuplexConnection {
                 }
             });
         };
+    }
+
+    @Override
+    public final Observable<Frame> getInput() {
+        return o -> {
+            o.onSubscribe(() -> subjects.removeIf(s -> s == o));
+            subjects.add(o);
+        };
+    }
+
+    @Override
+    public void addOutput(Publisher<Frame> o, Completable callback) {
+        o.subscribe(new Subscriber<Frame>() {
+            private Subscription subscription;
+
+            @Override
+            public void onSubscribe(Subscription s) {
+                subscription = s;
+                s.request(Long.MAX_VALUE);
+            }
+
+            @Override
+            public void onNext(Frame frame) {
+                try {
+                    ByteBuf byteBuf = Unpooled.wrappedBuffer(frame.getByteBuffer());
+                    BinaryWebSocketFrame binaryWebSocketFrame = new BinaryWebSocketFrame(byteBuf);
+                    ChannelFuture channelFuture = channel.writeAndFlush(binaryWebSocketFrame);
+                    channelFuture.addListener(future -> {
+                        Throwable cause = future.cause();
+                        if (cause != null) {
+                            if (cause instanceof ClosedChannelException) {
+                                onError(new TransportException(cause));
+                            } else {
+                                onError(cause);
+                            }
+                        }
+                    });
+                } catch (Throwable t) {
+                    onError(t);
+                }
+            }
+
+            @Override
+            public void onError(Throwable t) {
+                callback.error(t);
+                if (t instanceof TransportException) {
+                    subscription.cancel();
+                }
+            }
+
+            @Override
+            public void onComplete() {
+                callback.success();
+            }
+        });
+    }
+
+    @Override
+    public double availability() {
+        return channel.isOpen() ? 1.0 : 0.0;
+    }
+
+    @Override
+    public void close() throws IOException {
+        channel.close();
     }
 
     public String toString() {
@@ -103,5 +181,6 @@ public class ClientWebSocketDuplexConnection extends NettyDuplexConnection {
             + ", isRegistered=" + channel.isRegistered()
             + ", channelId=" + channel.id().asLongText()
             + "])";
+
     }
 }
